app/requirements.txt
Flask==2.0.1
Werkzeug==2.0
mysql-connector-python==8.3.0